+++ .github/actions/rust-build-release/tests/conftest.py
@@ -2,15 +2,18 @@
 
 from __future__ import annotations
 
+import typing as typ
+
 import pytest
+
+IteratorNone = typ.Iterator[None]
 
 
 @pytest.fixture
 def uncapture_if_verbose(
     request: pytest.FixtureRequest, capfd: pytest.CaptureFixture[str]
-):
+) -> IteratorNone:
     """Disable output capture when pytest runs with ``-v`` or higher verbosity."""
-
     if request.config.get_verbosity() > 0:
         with capfd.disabled():
             yield